Move socket to pkg
