+++ pkg/socket/interface.go
