+++ signal/coordinator/interface.go
@@ -1,15 +1,17 @@
 // Package coordinator contains handling socket more clearly
 package coordinator
 
-import "pdn/signal/controller/socket"
+import (
+	"pdn/pkg/socket"
+)
 
 // Coordinator is an interface for managing socket.
 //
 //go:generate mockgen -destination=mock_coordinator.go -package=coordinator . Coordinator
 type Coordinator interface {
-	RequestResponse(channelID string, userID string, data string) (string, error)
+	requestResponse(channelID string, userID string, data string) (string, error)
 	Activate(channelID string, userID string, s socket.Socket) error
-	Response(channelID, userID string, data string) error
+	response(channelID, userID string, data string) error
 	Remove(channelID, userID string) error
 	Send(channelID, userID, sdp string) (string, error)
 	Receive(channelID, userID, sdp string) (string, error)
